backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
aniso8601==10.0.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
auth0-python==4.7.1
beautifulsoup4==4.12.3
billiard==4.2.4
blinker==1.9.0
celery==5.4.0
certifi==2026.2.25
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==42.0.5
cuda-bindings==12.9.4
cuda-pathfinder==1.4.0
dataclasses-json==0.6.7
dnspython==2.8.0
faiss-cpu==1.9.0
fastapi==0.134.0
filelock==3.24.3
Flask==3.0.3
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.43
greenlet==3.3.2
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.6.2
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.3.31
langchain-core==1.2.16
langchain-huggingface==1.2.0
langchain-text-splitters==1.1.1
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
orjson==3.11.7
ormsgpack==1.12.2
packaging==24.2
pdf2image==1.17.0
pillow==12.1.1
prompt_toolkit==3.0.52
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.9.0
pymongo==4.10.1
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
redis==5.2.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==2.7.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
soupsieve==2.8.3
SQLAlchemy==2.0.47
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.10.0
tqdm==4.67.3
transformers==4.40.1
triton==3.6.0
typer==0.24.1
typer-slim==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.2.0
uuid_utils==0.14.1
uvicorn==0.41.0
vine==5.1.0
wcwidth==0.6.0
Werkzeug==3.1.6
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
PyMuPDF==1.27.1
